[Refactor] Split work group manager from work group
