+++ be/src/agent/resource_group_usage_recorder.cpp
@@ -17,6 +17,7 @@
 #include "base/time/time.h"
 #include "common/system/cpu_info.h"
 #include "exec/workgroup/work_group.h"
+#include "exec/workgroup/work_group_manager.h"
 #include "runtime/exec_env.h"
 
 namespace starrocks {
