+++ be/src/exec/aggregate/aggregate_base_node.cpp
@@ -15,6 +15,7 @@
 #include "exec/aggregate/aggregate_base_node.h"
 
 #include "exec/aggregator.h"
+#include "exec/runtime_filter/runtime_filter_descriptor.h"
 #include "exprs/expr_factory.h"
 
 namespace starrocks {
