+++ be/src/exec/pipeline/aggregate/aggregate_blocking_sink_operator.h
@@ -21,6 +21,7 @@
 #include "base/concurrency/race_detect.h"
 #include "exec/aggregator.h"
 #include "exec/pipeline/operator_factory.h"
+#include "exec/pipeline/runtime_filter_hub.h"
 #include "runtime/runtime_state_fwd.h"
 
 namespace starrocks::pipeline {
