+++ be/src/exec/pipeline/fragment_context.cpp
@@ -25,6 +25,7 @@
 #include "base/uid_util.h"
 #include "common/config_exec_flow_fwd.h"
 #include "common/logging.h"
+#include "common/thread/priority_thread_pool.hpp"
 #include "common/thread/threadpool.h"
 #include "common/util/thrift_client_cache.h"
 #include "compute_env/data_stream/data_stream_mgr.h"
@@ -38,6 +39,7 @@
 #include "exec/pipeline/schedule/timeout_tasks.h"
 #include "exec/workgroup/pipeline_executor_set.h"
 #include "exec/workgroup/work_group.h"
+#include "exec/workgroup/work_group_manager.h"
 #include "platform/thrift_rpc_helper.h"
 #include "runtime/batch_write/batch_write_mgr.h"
 #include "runtime/exec_env.h"
