+++ be/src/exec/pipeline/pipeline.cpp
@@ -21,6 +21,7 @@
 #include "exec/pipeline/pipeline_driver.h"
 #include "exec/pipeline/scan/connector_scan_operator.h"
 #include "exec/pipeline/scan/schema_scan_operator.h"
+#include "exec/workgroup/pipeline_executor_set.h"
 #include "runtime/runtime_state.h"
 
 namespace starrocks::pipeline {
