+++ be/src/exec/pipeline/pipeline_driver_executor.cpp
@@ -20,6 +20,7 @@
 #include "base/utility/defer_op.h"
 #include "common/config_exec_flow_fwd.h"
 #include "common/system/master_info.h"
+#include "common/thread/thread.h"
 #include "exec/pipeline/fragment_context.h"
 #include "exec/pipeline/pipeline_metrics.h"
 #include "exec/pipeline/query_context.h"
