+++ be/src/exec/pipeline/sink/memory_scratch_sink_operator.cpp
@@ -17,6 +17,7 @@
 #include "compute_env/result/result_queue_mgr.h"
 #include "exec/pipeline/fragment_context.h"
 #include "exec/pipeline/pipeline_driver_executor.h"
+#include "exec/workgroup/pipeline_executor_set.h"
 #include "exec/workgroup/work_group.h"
 #include "exprs/expr_executor.h"
 #include "exprs/expr_factory.h"
