+++ be/src/exec/workgroup/scan_executor.cpp
@@ -14,6 +14,7 @@
 
 #include "exec/workgroup/scan_executor.h"
 
+#include "common/thread/thread.h"
 #include "exec/pipeline/pipeline_metrics.h"
 #include "exec/workgroup/scan_task_queue.h"
 #include "runtime/current_thread.h"
