+++ be/src/exec/workgroup/scan_task_queue.cpp
@@ -18,6 +18,7 @@
 #include "common/status.h"
 #include "exec/workgroup/work_group.h"
 #include "exec/workgroup/work_group_fwd.h"
+#include "exec/workgroup/work_group_manager.h"
 #include "runtime/exec_env.h"
 
 namespace starrocks::workgroup {
