+++ be/src/exec/workgroup/work_group.h
@@ -15,25 +15,23 @@
 #pragma once
 
 #include <atomic>
+#include <chrono>
+#include <cstddef>
+#include <cstdint>
 #include <memory>
-#include <mutex>
-#include <queue>
-#include <unordered_map>
+#include <optional>
+#include <string>
+#include <utility>
 
 #include "base/time/time.h"
-#include "common/thread/priority_thread_pool.hpp"
-#include "exec/pipeline/pipeline_driver_queue.h"
+#include "base/types/int128.h"
+#include "common/statusor.h"
 #include "exec/pipeline/pipeline_fwd.h"
 #include "exec/workgroup/work_group_fwd.h"
-#include "mem_tracker_manager.h"
-#include "pipeline_executor_set_manager.h"
+#include "gen_cpp/WorkGroup_types.h"
 #include "runtime/mem_tracker.h"
-#include "storage/olap_define.h"
 
 namespace starrocks {
-
-class MetricRegistry;
-class TWorkGroup;
 
 namespace workgroup {
 
@@ -41,8 +39,6 @@ using steady_clock = std::chrono::steady_clock;
 
 using pipeline::QueryContext;
 using WorkGroupType = TWorkGroupType::type;
-struct WorkGroupMetrics;
-using WorkGroupMetricsPtr = std::shared_ptr<WorkGroupMetrics>;
 
 template <typename Q>
 class WorkGroupSchedEntity {
@@ -117,7 +113,7 @@ public:
     WorkGroup(std::string name, int64_t id, int64_t version, size_t cpu_weight, double memory_limit, size_t concurrency,
               double spill_mem_limit_threshold, WorkGroupType type, std::string mem_pool);
     explicit WorkGroup(const TWorkGroup& twg);
-    ~WorkGroup() = default;
+    ~WorkGroup();
 
     void init(std::shared_ptr<MemTracker>& parent_mem_tracker);
 
@@ -206,18 +202,15 @@ public:
     int64_t cpu_runtime_ns() const { return _cpu_runtime_ns; }
     std::string mem_pool() const { return _mem_pool; }
     void set_shared_executors(PipelineExecutorSet* executors) { _executors = executors; }
-    void set_exclusive_executors(std::unique_ptr<PipelineExecutorSet> executors) {
-        _exclusive_executors = std::move(executors);
-        _executors = _exclusive_executors.get();
-    }
+    void set_exclusive_executors(std::unique_ptr<PipelineExecutorSet> executors);
 
     PipelineExecutorSet* exclusive_executors() const { return _exclusive_executors.get(); }
     PipelineExecutorSet* executors() const { return _executors; }
 
-    static constexpr int64 DEFAULT_WG_ID = 0;
-    static constexpr int64 DEFAULT_MV_WG_ID = 1;
-    static constexpr int64 DEFAULT_VERSION = 0;
-    static constexpr int64 DEFAULT_MV_VERSION = 1;
+    static constexpr int64_t DEFAULT_WG_ID = 0;
+    static constexpr int64_t DEFAULT_MV_WG_ID = 1;
+    static constexpr int64_t DEFAULT_VERSION = 0;
+    static constexpr int64_t DEFAULT_MV_VERSION = 1;
     inline static std::string DEFAULT_MEM_POOL{"default_mem_pool"};
 
     // Yield scan io task when maximum time in nano-seconds has spent in current execution round.
@@ -274,95 +267,6 @@ private:
 
     std::unique_ptr<PipelineExecutorSet> _exclusive_executors;
     PipelineExecutorSet* _executors = nullptr;
-};
-
-// WorkGroupManager is a singleton used to manage WorkGroup instances in BE, it has an io queue and a cpu queues for
-// pick next workgroup for computation and launching io tasks.
-class WorkGroupManager {
-public:
-    explicit WorkGroupManager(PipelineExecutorSetConfig executors_manager_conf, MetricRegistry* metrics = nullptr);
-
-    ~WorkGroupManager();
-
-    Status start();
-
-    // add a new workgroup to WorkGroupManger
-    WorkGroupPtr add_workgroup(const WorkGroupPtr& wg);
-    // return reserved beforehand default workgroup for query is not bound to any workgroup
-    WorkGroupPtr get_default_workgroup();
-    // return reserved beforehand default mv workgroup for MV query is not bound to any workgroup
-    WorkGroupPtr get_default_mv_workgroup();
-
-    size_t num_workgroups() const { return _workgroups.size(); }
-
-    void close();
-    // destruct workgroups
-    void destroy();
-
-    void apply(const std::vector<TWorkGroupOp>& ops);
-    std::vector<TWorkGroup> list_workgroups();
-    std::vector<std::string> list_memory_pools() const;
-
-    using WorkGroupConsumer = std::function<void(const WorkGroup&)>;
-    void for_each_workgroup(const WorkGroupConsumer& consumer) const;
-
-    void update_metrics();
-
-    bool should_yield(const WorkGroup* wg) const;
-    PipelineExecutorSet* shared_executors() const { return _executors_manager.shared_executors(); }
-    void for_each_executors(const ExecutorsManager::ExecutorsConsumer& consumer) const;
-    void change_num_connector_scan_threads(uint32_t num_connector_scan_threads);
-    void change_enable_resource_group_cpu_borrowing(bool val);
-    void change_exec_state_report_max_threads(int max_threads);
-    void change_priority_exec_state_report_max_threads(int max_threads);
-    void set_workgroup_expiration_time(std::chrono::seconds value);
-
-private:
-    using MutexType = std::shared_mutex;
-    using UniqueLockType = std::unique_lock<MutexType>;
-    using SharedLockType = std::shared_lock<MutexType>;
-
-    // {create, alter,delete}_workgroup_unlocked is used to replay WorkGroupOps.
-    // WorkGroupManager::_mutex is held when invoking these method.
-    void create_workgroup_unlocked(const WorkGroupPtr& wg, UniqueLockType& lock);
-    void alter_workgroup_unlocked(const WorkGroupPtr& wg, UniqueLockType& lock);
-    void delete_workgroup_unlocked(const WorkGroupPtr& wg);
-    void add_metrics_unlocked(const WorkGroupPtr& wg, UniqueLockType& unique_lock);
-    void update_metrics_unlocked();
-    WorkGroupPtr get_default_workgroup_unlocked();
-
-private:
-    friend class ExecutorsManager;
-
-    mutable std::shared_mutex _mutex;
-    // Place it before _workgroups to ensure the shared executors is destructed after all the dedicated executors for
-    // workgroups, since _executors_manager owns the shared executors, and WorkGroup owns the dedicated executors.
-    ExecutorsManager _executors_manager;
-
-    std::unordered_map<int128_t, WorkGroupPtr> _workgroups;
-    std::unordered_map<int64_t, int64_t> _workgroup_versions;
-    std::list<int128_t> _workgroup_expired_versions;
-    std::chrono::seconds _workgroup_expiration_time{120};
-
-    std::atomic<size_t> _sum_cpu_weight = 0;
-    MetricRegistry* _metrics = nullptr;
-    MemTrackerManager _shared_mem_tracker_manager;
-    std::once_flag init_metrics_once_flag;
-    std::unordered_map<std::string, WorkGroupMetricsPtr> _wg_metrics;
-};
-
-class DefaultWorkGroupInitialization {
-public:
-    DefaultWorkGroupInitialization(WorkGroupManager* workgroup_manager, int64_t max_executor_threads);
-
-    // create or renew default group
-    std::shared_ptr<WorkGroup> create_default_workgroup();
-    // create or renew default mv group
-    std::shared_ptr<WorkGroup> create_default_mv_workgroup();
-
-private:
-    WorkGroupManager* _workgroup_manager;
-    int64_t _max_executor_threads;
 };
 
 } // namespace workgroup
