+++ be/src/service/service_be/config_update_hooks.cpp
@@ -43,6 +43,7 @@
 #include "common/thread/priority_thread_pool.hpp"
 #include "common/util/bthreads/executor.h"
 #include "exec/workgroup/scan_executor.h"
+#include "exec/workgroup/work_group_manager.h"
 #include "runtime/batch_write/batch_write_mgr.h"
 #include "runtime/batch_write/txn_state_cache.h"
 #include "runtime/env/global_env.h"
