+++ be/test/exec/pipeline/pipeline_driver_queue_test.cpp
@@ -23,6 +23,7 @@
 #include "exec/pipeline/pipeline_metrics.h"
 #include "exec/pipeline/query_context.h"
 #include "exec/workgroup/work_group.h"
+#include "exec/workgroup/work_group_manager.h"
 #include "runtime/exec_env.h"
 
 namespace starrocks::pipeline {
