+++ be/test/exec/workgroup/lock_free_work_group_scan_task_queue_test.cpp
@@ -23,6 +23,7 @@
 #include "base/concurrency/countdown_latch.h"
 #include "common/thread/thread.h"
 #include "exec/workgroup/work_group.h"
+#include "exec/workgroup/work_group_manager.h"
 #include "runtime/exec_env.h"
 
 namespace starrocks::workgroup {
